rag/pyproject.toml
[project]
name = "emergency-healthcare-rag"
version = "0.1.0"
description = "Emergency Healthcare RAG API"
requires-python = ">=3.11"
dependencies = [
    # API dependencies
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic==2.4.2",
    "loguru>=0.7.0",
    # Core dependencies
    "requests>=2.31.0",
    "numpy",
    "ollama\u003e=0.1.0",
    "google-genai\u003e=0.8.0",
    "httpx>=0.25.0",
    "httpcore>=1.0.0",
    "h11>=0.12.0",
    # Embeddings dependencies
    "sentence-transformers>=2.2.0",
    "faiss-cpu>=1.7.4",
    "torch>=2.0.0",
    "transformers>=4.0.0",
    "tqdm>=4.62.3",
    "PyYAML>=6.0",
    "huggingface-hub>=0.20.3",
    "urllib3>=2.0.0",
    "idna>=3.4",
    "charset-normalizer>=3.1.0",
    "certifi>=2024.7.4",
    "typing-extensions>=4.7.1",
    "filelock>=3.12.2",
    "scipy>=1.10.1",
    "scikit-learn>=1.3.0",
    "joblib>=1.2.0",
    "threadpoolctl>=2.2.0",
    "packaging>=23.0",
    "regex>=2023.10.3",
    "safetensors>=0.4.3",
    "tokenizers>=0.19.1",
    "sympy>=1.12",
    "mpmath>=1.3.0",
    "Pillow>=9.5.0",
    "anyio>=4.0.0",
    "sniffio>=1.3.0",
    "pydantic-core==2.10.1",
    "python-dotenv>=1.0.0",
    "typing-inspection>=0.4.0",
    "annotated-types>=0.6.0",
    "bm25s>=0.2.13",
    "pystemmer>=3.0.0",
    "accelerate>=0.20.0",
    "anthropic>=0.60.0",
    "sentence-transformers>=5.0.0",
    "pandas>=2.3.1",
    "datasets>=4.0.0",
    "ragas>=0.1.0",
    "plotly>=5.0.0",
    "chromadb>=1.0.15",
]

[project.scripts]
api = "api:start_server"
validate = "validate:main"
run-rag = "run_rag:main"

[tool.healthcare-rag]
default_embedding = "pubmedbert-base-embeddings"
default_llm = "cogito:8b"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]
